# docs/06-api/openapi.yaml
openapi: 3.1.0
info:
  title: AI BIOS Logo Platform API
  version: 1.0.0
  description: Enterprise API for modernist logo design intelligence
servers:
  - url: http://localhost:3001/api
    description: Development
paths:
...
  /prompts/generate:
    post:
      summary: Generate ranked prompt variations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [industry]
              properties:
                industry: { type: string }
                companyName: { type: string }
                variationCount: { type: integer, minimum: 1, maximum: 100 }
                inspirationMode: { type: string }
                minimalismLevel: { type: integer, minimum: 1, maximum: 10 }
      responses:
        '200':
          description: Generated prompts with recommendations
  /prompts/recommend/{industry}:
    get:
      summary: Get principle recommendations for industry
      parameters:
        - name: industry
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Recommendations and suggested principles
...